Fix/hack 1. ========
At present, if you wish to load a .hhc file instead of add books manually, you can only have one book - UNLESS you omit the outer <ul> </ul>. Then the contents will appear to have more than book, although in fact it's just one book that's displayed on level zero and therefore get the Book icons.

However, you'll also get a dummy 'book' with the name of the project but no children. So my hack detects whether the .hhc file is trying to add books at level zero, and if so, throws away the first 'book', so that only the hhc-specified books get shown.

So the end result is that our hhc file can specify multiple sections that don't all hang off one dummy book node which itself hangs off the root node -- which is a UI pain!

Fix 2.
====

I changed  the test

        if (!m_Page.IsEmpty())

to

        if (tag.GetParam("TYPE") == "text/sitemap")

at around L.120 of helpdata.cpp as suggested in the comment, so we don't have to specify a URL for the node to be valid. Sometimes you don't want a URL in a contents item. I've added some tests to stop it complaining if you click on a URL-less item.

Feature enhancement/Hack 3.
======================

I've added wxHtmlHelpController::SetBookBasePath so that the .hhp/.hhc files don't have to be in the same hierarchy as the HTML files.


git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@9792 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775