Applied patch [ 708582 ] wxGenericFileDialog & expose generic wxFileCtrl By John Labenski (jrl1)
The wxFileCtrl is a nice little widget that when
married with a wxDirCtrl makes for a simple, easy to
use explorer type widget.

include/wx/generic/filedlgg.h, src/generic/filedlg.cpp

Expose wxFileData and wxFileCtrl, rename wxFileDialog
to wxGenericFileDialog, it can be used in MSW now.

wxFileCtrl - added sorting for clicking in the col headers.
let wxGenericFileDialog manage it's own static text.
allow for viewing of root (for showing drives) in MSW
wxFileData - moved enum for wxFileCtrl col's into it.
added IsDrive and use bitmask instead of a slew of
bools for IsDir...

A couple bug fixes for wxFileCtrl for MSW, c:\.. is not
a file and can't be stated, show drives.

Basicly its the same as before, with no change besides
sorting for GTK. For MSW, the wxFileCtrl works in a few
more cases, hopefully all the MGL etc code was
untouched and should still work.


git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@19964 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775